test: split Passage tests by directive category
